pyproject.toml
[project]
name = "kibo-core"
version = "0.1.0"
description = "Universal Distributed Agent Framework"
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "ray",
    "pydantic",
    "langchain",
    "langgraph",
    "langchain-ollama",
    "crewai>=0.80.0",
    "langchain-openai",
    "agno>=2.4.8",
    "pydantic-ai>=0.4.3",
    "litellm[proxy]>=1.61.0",
    "duckduckgo-search<7.0.0",
    "langchain-community>=0.3.31",
    "crewai-tools>=0.76.0",
    "tavily-python>=0.7.21",
    "python-dotenv>=1.1.1",
]

[project.scripts]
kibo = "kibo_core.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kibo_core"]
